+++ lib/td_dd/data_structures/loader_worker.ex
@@ -6,6 +6,7 @@ defmodule TdDd.Loader.LoaderWorker do
   use GenServer
 
   alias TdCache.TaxonomyCache
+  alias TdDd.Cache.StructureLoader
   alias TdDd.CSV.Reader
   alias TdDd.DataStructures.Ancestry
   alias TdDd.Loader
@@ -13,8 +14,6 @@ defmodule TdDd.Loader.LoaderWorker do
   alias TdDd.Systems
 
   require Logger
-
-  @index_worker Application.get_env(:td_dd, :index_worker)
 
   @structure_import_schema Application.get_env(:td_dd, :metadata)[:structure_import_schema]
   @structure_import_required Application.get_env(:td_dd, :metadata)[:structure_import_required]
@@ -227,7 +226,7 @@ defmodule TdDd.Loader.LoaderWorker do
   defp do_post_process(data_structure_ids, nil) do
     # If any ids have been returned by the bulk load process, these
     # data structures should be reindexed.
-    @index_worker.reindex(data_structure_ids)
+    StructureLoader.refresh(data_structure_ids)
   end
 
   defp do_post_process(data_structure_ids, external_id) do